Add files via upload
This code can now run outside the directory which was previously required to run in the directory of the data. The final output now contains the full catalog information that is extracted from pybdsf along with spectral index and observed source location which will be helpful for users to get complete information about the source. The catalog has not been arranged properly but individual columns can be accessed by the user to gain additional source information. 
Please not that now one only needs to give correct paths to the required directory however currently for our purpose I have used region as an indicator for bookkeeping purposes, So it is expected that one keeps the images in a directory labeled by its region eg : /home/work/DATA/SEARCH/G033.0-5.0/G033.0-5.0.FITS