Add files via upload
The current code uses the CSV file of the pulsar candidates generated by SIBPF to check the variability at a particular pixel location. The SIBPF code gives the X and Y co ordinates of the sources . The sources with same X and Y co ordinates measured at different days are then written to a separate file which will contain the flux values of each observation as a separate column and the column names are edited appropriately. The code calculated the reduced ch-square statistics assuming no variability , this has be referenced from Bell et.al and Sett et.al. 
Currently no crossmatching is done however I am trying to also figure that out if we can use pybdsf on specific pixel locations instead of the whole image. 
The code has been tested for an image and its copy. If I could get any images for other days i could try testing it